Combined __init__.py and core.py into lcn.py component. Bumped to pypck==0.5.6
