+++ requirements_all.txt
@@ -1115,7 +1115,7 @@ pyotp==2.2.6
 pyowm==2.9.0
 
 # homeassistant.components.lcn
-pypck==0.5.5
+pypck==0.5.6
 
 # homeassistant.components.media_player.pjlink
 pypjlink2==1.2.0
